backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.1
django-cors-headers
djangorestframework-simplejwt
channels==4.1.0
daphne==4.1.2
channels-redis==4.2.0



















